Added sql lite, and changed the app to read music in the internal storage

+++ app/src/main/java/com/Jamie/juicewrldmusicplayer/Song.kt
@@ -1,14 +1,13 @@
 package com.Jamie.juicewrldmusicplayer
 
-
-
+import android.net.Uri
 
 
 data class Song(
     val id:Int,
     val songName: String,
     val albumName: String,
-    val albumImage: Int,
+    val albumImage: Long,
     val audioFile: String,
     val amountPlayed: Int
 )
